requirements.txt
fastapi==0.98.0
uvicorn==0.23.0
tensorflow==2.10.0
pillow==9.2.0
numpy==1.23.0
opencv-python==4.10.0.84
python-multipart
inference-sdk
